+++ lizard_portal/media/lizard_portal/lizard/grid/EditableGrid.js
@@ -69,7 +69,10 @@
       addEditIcon: false,
       addDeleteIcon: false,
       actionEditIcon: null,
-      actionDeleteIcon: null
+      actionDeleteIcon: null,
+      usePagination: true,
+      read_only_field: null,
+      recordsPerPage: 25
     },
     extraEditors: {
       timeserie: {
@@ -136,9 +139,12 @@
         }
       }
     },
-    get_editor: function(col) {
+    get_editor: function(record, col) {
       var editor, me, type;
       me = this;
+      if (me.read_only_field && record.data[me.read_only_field] === true) {
+        return false;
+      }
       if (typeof col.editable === 'undefined') {
         col.editable = true;
       }
@@ -312,7 +318,7 @@
         };
         if (col.editable) {
           col_config.getEditor = Ext.Function.bind(function(record, col) {
-            return me.get_editor(col);
+            return me.get_editor(record, col);
           }, me, [col], true);
         }
         if (col.editIf) {
@@ -400,7 +406,11 @@
       var selection;
       selection = this.getView().getSelectionModel().getSelection()[0];
       if (selection) {
-        return this.store.remove(selection);
+        if (this.read_only_field && selection.data[this.read_only_field] === true) {
+          return false;
+        } else {
+          return this.store.remove(selection);
+        }
       }
     },
     getStoreConfig: function() {
@@ -446,6 +456,7 @@
       store = {
         type: 'leditstore',
         fields: fields,
+        pageSize: this.recordsPerPage,
         proxy: {
           type: 'ajax',
           api: {
@@ -467,8 +478,7 @@
             encode: true,
             successProperty: 'success',
             totalProperty: 'count'
-          },
-          autoLoad: true
+          }
         }
       };
       return store;
@@ -476,6 +486,9 @@
     initComponent: function() {
       var me;
       me = this;
+      if (!this.getUsePagination()) {
+        this.recordsPerPage = 10000;
+      }
       me.columns = this.getColumnConfig();
       me.store = Ext.create('Lizard.store.EditGridStore', this.getStoreConfig());
       me.bbar = [];
@@ -539,13 +552,15 @@
           }
         ]);
       }
-      this.bbar = {
-        xtype: 'pagingtoolbar',
-        pageSize: 25,
-        store: me.store,
-        displayInfo: true,
-        items: ['-'].concat(me.bbar)
-      };
+      if (this.getUsePagination()) {
+        this.bbar = {
+          xtype: 'pagingtoolbar',
+          pageSize: this.recordsPerPage,
+          store: me.store,
+          displayInfo: true,
+          items: ['-'].concat(me.bbar)
+        };
+      }
       return this.callParent(arguments);
     }
   });