+++ lizard_portal/templates/portals/krw-overzicht.js
@@ -29,10 +29,8 @@
 		items: [{
             title: 'Grafieken',
             flex: 1,
-            xtype: 'multigraph',
-            graph_service_url: '/graph/',
-            context_manager: Ext.getCmp('portalWindow').context_manager,
-            graphs: {% get_portal_template graphs-krw-overzicht %}
+            xtype: 'multigraphstore',
+            store: Ext.create('Lizard.store.Graph', {data: {% get_portal_template graphs-krw-overzicht %} })
 		}]
 	}]
 }