+++ lizard_portal/templates/portals/maatregelen.js
@@ -25,13 +25,45 @@
                 var url = '/measure/summary/'+ cm.object_id +'/krw_measures/';
                 me.loader.load({
                     url:url,
-                    method: 'GET'
+                    method: 'GET',
+                    success: function() {
+                      reloadGraphs();
+                      me.setLoading(false);
+                    },
+                    failure: function() {
+                      me.setLoading(false);
+                    }
                 });
-                me.setLoading(false);
             },
             loader:{
                 renderer: 'html'
-            }
+            },
+            tools: [{
+                type: 'plus',
+                handler: function (e, target, panelHeader, tool) {
+                    Ext.create('Ext.window.Window', {
+                        title: 'Nieuwe maatregel toevoegen',
+                        width: 800,
+                        height: 600,
+                        modal: true,
+                        finish_edit_function: function (updated_record) {
+                            //todo
+                        },
+                        editpopup: true,
+                        loader:{
+                            loadMask: true,
+                            autoLoad: true,
+                            url: '/measure/measure_detailedit_portal/',
+                            ajaxOptions: {
+                                method: 'GET'
+                            },
+                            renderer: 'component'
+                        }
+                    }).show();
+                }
+            } ]
+
+
 		}]
 	}]
 }