+++ lizard_portal/templates/portals/toestand-aan-afvoergebied.js
@@ -5,7 +5,7 @@
     breadcrumbs: [
         {
             name: 'watersysteemkaart',
-            link: 'watersysteemkaart'
+            link: 'homepage'
         },
         {
             name: 'Toestand'
@@ -25,14 +25,11 @@
     },{
 		flex: 1,
 		items: [{
-			title: 'Grafieken',
+            title: 'Grafieken',
             flex: 1,
-            xtype: 'multigraph',
-            graph_service_url: '/graph/',
-            context_manager: Ext.getCmp('portalWindow').context_manager,
-            graphs: {% get_portal_template graphs-krw-overzicht %}
+            xtype: 'multigraphstore',
+            store: Ext.create('Lizard.store.Graph', {data: {% get_portal_template graphs-aanafvoer_toestand %} })
 		}]
-
     },
     {
 		width: 200,