+++ lizard_portal/templates/portals/waterbalans.js
@@ -21,8 +21,34 @@
 	items:[{
 		width: 200,
 		items: [{
-			title: 'Samenvatting instellingen',
-            flex: 1,
+			title: 'Maatregelen',
+            flex:1,
+            autoScroll: true,
+            plugins: [
+                'applycontext'
+            ],
+            applyParams: function(params) {
+                var me = this;
+                me.setLoading(true);
+                var cm = Ext.getCmp('portalWindow').context_manager.getContext();
+
+                me.loader.load({
+                    url: 'wbconfiguration/api/summary/',
+                    params: {
+                        object_id: cm.object_id
+                    },
+                    method: 'GET',
+                    success: function() {
+                      me.setLoading(false);
+                    },
+                    failure: function() {
+                      me.setLoading(false);
+                    }
+                });
+            },
+            loader:{
+                renderer: 'html'
+            },
             bbar: [{
                 xtype: 'button',
                 text: 'Details',
@@ -42,13 +68,10 @@
 	},{
 		flex: 1,
 		items: [{
-			title: 'Grafieken',
+            title: 'Grafieken',
             flex: 1,
-            xtype: 'multigraph',
-            graph_service_url: '/graph/',
-            context_manager: Ext.getCmp('portalWindow').context_manager,
-            graphs: {% get_portal_template graphs-waterbalans %}
-
+            xtype: 'multigraphstore',
+            store: Ext.create('Lizard.store.Graph', {data: {% get_portal_template graphs-waterbalans %} })
 		}]
 	}]
 }